# api/openapi/workstream-gateway.yaml
openapi: 3.1.0

info:
  title: Workstream Gateway API
  description: |
    Public API gateway for the Workstream platform, routing client requests to other services.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server.

paths:
  /tasks:
    get:
      operationId: getTasks
      summary: List existing tasks
      tags: [Tasks]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: A list of tasks
          content:
            application/json:
              schema:
                type: object
                required: [tasks]
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
              example:
                tasks:
                  - id: "0b2f3e9a-7d1b-4b4c-9c2b-7b9e8ef1e1a1"
                    title: "Buy milk"
                    completed: false
                  - id: "a7c4c2c4-6c4d-4b1b-8f34-3b9a4a6c2b77"
                    title: "Call the dentist"
                    completed: true
        "401":
          description: Missing or invalid JWT
        "403":
          description: Authenticated but not authorized

    post:
      operationId: createTask
      summary: Creates a task
      tags: [Tasks]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
            example:
              title: Buy milk
      responses:
        "201":
          description: Task successfully created
          content:
            application/json:
              schema:
                type: object
                required: [task]
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
              example:
                task:
                  id: "0b2f3e9a-7d1b-4b4c-9c2b-7b9e8ef1e1a1"
                  title: "Buy milk"
                  completed: false
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "title is required"
        "401":
          description: Missing or invalid JWT
        "403":
          description: Authenticated but not authorized

  /tasks/{id}:
    get:
      operationId: getTask
      summary: Returns an existing task
      tags: [Tasks]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: A task
          content:
            application/json:
              schema:
                type: object
                required: [task]
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
              example:
                task:
                  id: "0b2f3e9a-7d1b-4b4c-9c2b-7b9e8ef1e1a1"
                  title: "Buy milk"
                  completed: false
        "401":
          description: Missing or invalid JWT
        "403":
          description: Authenticated but not authorized
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "task not found"

  /tasks/{id}/complete:
    post:
      operationId: completeTask
      summary: Completes a task
      tags: [Tasks]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task successfully completed
        "401":
          description: Missing or invalid JWT
        "403":
          description: Authenticated but not authorized
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "task not found"

  /identity/register:
    post:
      operationId: register
      summary: Register new user
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - roles
                - organizationId
              properties:
                email:
                  type: string
                  format: email
                organizationId:
                  type: string
                  format: uuid
                password:
                  type: string
                  minLength: 6
                roles:
                  type: array
                  items:
                    type: string
                    enum: [user, admin]
            example:
              email: test2@mail.com
              password: test2-pass
              roles: ["admin"]
      responses:
        "201":
          description: User successfully registered
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token:
                    type: string
              example:
                token: eyJhbGciOi...

  /identity/login:
    post:
      operationId: login
      summary: Login existing user
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
            example:
              email: test2@mail.com
              password: test2-pass
      responses:
        "200":
          description: Successfully logged in
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token:
                    type: string
              example:
                token: eyJhbGciOi...

  /identity/users/{id}/unregister:
    patch:
      operationId: unregisterUser
      summary: Unregister user
      tags: [Identity, Identity/Admin]
      description: |
        Unregisters a user.

        Authorization rules:
        - The authenticated user may unregister their own account.
        - An authenticated user with the `admin` role may unregister any user.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - BearerAuth: []
      responses:
        "204":
          description: user unregistered
        "401":
          description: unauthorized
        "403":
          description: forbidden

  /identity/request-password:
    post:
      operationId: requestPassword
      summary: Request user password reset
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
              example:
                email: test2@mail.com
      responses:
        "204":
          description: password reset requested

  /identity/reset-password:
    post:
      operationId: resetPassword
      summary: Reset user password
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reset_token
                - password
              properties:
                reset_token:
                  type: string
                password:
                  type: string
                  minLength: 6
              example:
                reset_token: eyJhbGciOi...
                password: test-pass
      responses:
        "204":
          description: password reset requested

  /identity/users:
    get:
      tags: [Identity/Admin]
      summary: List users (admin only)
      operationId: listAllUsers
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Missing or invalid JWT
        "403":
          description: Authenticated but not authorized (not admin)

  /identity/users/{id}:
    get:
      tags: [Identity/Admin]
      summary: Get user by ID (admin only)
      description: |
        Returns a user
      operationId: getUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Missing or invalid JWT
        "403":
          description: Authenticated but not authorized (not admin)
        "404":
          description: User not found

    put:
      tags: [Identity/Admin]
      summary: Update user (admin only)
      description: |
        Updates a user
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated successfully
        "400":
          description: Invalid input
        "401":
          description: Missing or invalid JWT
        "403":
          description: Authenticated but not authorized (not admin)
        "404":
          description: User not found
        "409":
          description: Email already exists

    delete:
      tags: [Identity/Admin]
      summary: Delete user (admin only)
      description: |
        Deletes a user
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User deleted successfully
        "401":
          description: Missing or invalid JWT
        "403":
          description: Authenticated but not authorized (not admin)
        "404":
          description: User not found

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Task:
      type: object
      required: [id, title, completed]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        completed:
          type: boolean

    User:
      type: object
      required: [id, email, roles, isActive, organizationId]
      properties:
        id:
          type: string
          format: uuid
          example: 37ef199c-e2c6-422d-9195-d166c313390f
        email:
          type: string
          format: email
          example: admin@acme.com
        roles:
          type: array
          description: Roles assigned to the user within the organization.
          items:
            type: string
            enum: [admin, user]
        isActive:
          type: boolean
          example: true
        organizationId:
          type: string
          format: uuid

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        roles:
          type: array
          items:
            type: string
            enum: [admin, user]
        organizationId:
          type: string
          format: uuid
        isActive:
          type: boolean